repair: extract hashers
